repliker dagerPerUke fra deltakelsesmengder
Vises også i frontend for enkeltplassene

+++ mulighetsrommet-api/server/src/main/kotlin/no/nav/mulighetsrommet/api/utbetaling/model/UtbetalingBeregningFastSatsPerBenyttetPlassPerManed.kt
@@ -80,7 +80,7 @@ object FastSatsPerTiltaksplassPerManedBeregning :
                     Periode.of(mengde.gyldigFra, gyldigTil)?.intersect(periode)?.let { overlappingPeriode ->
                         DeltakelsesprosentPeriode(
                             periode = overlappingPeriode,
-                            deltakelsesprosent = mengde.deltakelsesprosent,
+                            deltakelsesprosent = mengde.deltakelsesprosent.toDouble(),
                         )
                     }
                 }
